Revert "remove warnings (merge request !302)"
This reverts commit 43103c809894cbc635aa32b3035df4afcd6ee03c.

+++ 3rdparty/LLM_kernels/csrc/kernels/nvidia/samplers/samplers_test.cu
@@ -253,7 +253,7 @@ TEST_F(LlamaNvidiaSamplersTestSuit, InvokeRepetitionPenaltyTest) {
     DataType* logits_ptr = reinterpret_cast<DataType*>(logits_host.data_ptr);
     DataType* repetition_penalties_ptr = reinterpret_cast<DataType*>(repetition_penalties_host.data_ptr);
     DataType* output_ptr = reinterpret_cast<DataType*>(output_ref.data_ptr);
-    for (size_t i = 0; i < vocab_size; ++i) {
+    for (int i = 0; i < vocab_size; ++i) {
       output_ptr[i] = logits_ptr[i] > 0 ? (logits_ptr[i] / repetition_penalties_ptr[i])
                                         : (logits_ptr[i] * repetition_penalties_ptr[i]);
     }
